+++ server/plugins/glink/admin/index.vue
@@ -1,0 +1,7 @@
+<script setup>
+import GLink from './GLink.vue'
+</script>
+
+<template>
+  <GLink />
+</template>